backend/requirements.txt
flask
flask_cors
numpy
Pillow
matplotlib
requests
scipy
